Refactor: Overhaul script with modular architecture and enhancements
This commit introduces a comprehensive refactoring of the HelloAsso sales report script to improve its structure, performance, and robustness.

The main script, `HelloAssoOrderStats.py`, is now a high-level orchestrator. The core logic has been split into a `src` directory with the following modules:
- `config.py`: Centralizes configuration loading and validation.
- `api.py`: Handles all HelloAsso API interactions, including a new data caching mechanism to reduce API calls.
- `processing.py`: Contains all business logic for data transformation and calculation.
- `reporting.py`: Manages the generation of all outputs (CSV files, console logs, plots, and emails).

Key improvements:
- **Performance:** Eliminated the N+1 query problem by removing redundant API calls inside a loop.
- **Robustness:** Added validation for `config.ini` to ensure all required keys are present at startup. Implemented more specific error handling for API and SMTP exceptions.
- **Maintainability:** The new modular structure makes the code easier to read, debug, and extend.
